created simple sql query

+++ internal/domain/domain.go
@@ -3,7 +3,7 @@ package domain
 type User struct {
 	Id     int    `json:"id"`
 	Name   string `json:"name"`
-	ChatId int    `json:"chat_id"`
+	ChatId int64  `json:"chat_id"`
 }
 
 type Audio struct {
